Add initial draft for File resource on steroids
This draft enumerates and describes the parameters that would be used by a File resource that performs a wide range of functions,
such as mirroring a set of directories, handling file permissions, etc... After discussion, this is not the initial direction we will take.
Instead, we will favor a simpler approach for checking the existence of a file, and checking if a file contains a matching string or
content. The new direction will initially be a `get` only resource. Later iterations could enable regex comparison and delete.
Whether or not this resource would ever handle file permissions, etc... can be decided once an initial implementation is complete.