Add new draft of file resource
This version of the resource reduces the range of functions the resource can perform.
First it is limited to a read-only resource.
It also limits the resource to basically getting some minimal properties and an option for find-in-file. The find-in-file is the "like" property that works as a wildcard search.

This concept is being abandoned. The read-only limitation is too much.
This implementation doesn't lend well to read-write,
the write becomes cumbersome to specify and how to handle before/after is unclear as well.